Split the storage service into single-purpose utils
