+++ packages/twenty-front-component-renderer/src/host/utils/clearFrontComponentStorage.ts
@@ -1,0 +1,18 @@
+import { buildFrontComponentStorageKeyPrefix } from '@/host/utils/buildFrontComponentStorageKeyPrefix';
+import { getBrowserStorageForArea } from '@/host/utils/getBrowserStorageForArea';
+import { getBrowserStorageKeysWithPrefix } from '@/host/utils/getBrowserStorageKeysWithPrefix';
+import { type FrontComponentStorageScope } from '@/types/FrontComponentStorageScope';
+
+export const clearFrontComponentStorage = ({
+  area,
+  ...namespace
+}: FrontComponentStorageScope): void => {
+  const browserStorage = getBrowserStorageForArea(area);
+
+  for (const namespacedKey of getBrowserStorageKeysWithPrefix(
+    browserStorage,
+    buildFrontComponentStorageKeyPrefix(namespace),
+  )) {
+    browserStorage.removeItem(namespacedKey);
+  }
+};
