+++ packages/twenty-front-component-renderer/src/host/utils/getBrowserStorageForArea.ts
@@ -1,0 +1,5 @@
+import { type FrontComponentStorageArea } from 'twenty-sdk/front-component';
+
+export const getBrowserStorageForArea = (
+  area: FrontComponentStorageArea,
+): Storage => (area === 'local' ? window.localStorage : window.sessionStorage);
