+++ packages/twenty-front-component-renderer/src/index.ts
@@ -134,7 +134,9 @@ export { installStyleBridge } from './polyfills/installStyleBridge';
 export { exposeGlobals } from './remote/utils/exposeGlobals';
 export type { FrontComponentExecutionContext } from 'twenty-sdk/front-component';
 export type { FrontComponentHostCommunicationApi } from './types/FrontComponentHostCommunicationApi';
-export { frontComponentStorageService } from './host/utils/frontComponentStorageService';
+export { setFrontComponentStorageItem } from './host/utils/setFrontComponentStorageItem';
+export { deleteFrontComponentStorageItem } from './host/utils/deleteFrontComponentStorageItem';
+export { clearFrontComponentStorage } from './host/utils/clearFrontComponentStorage';
 export type { FrontComponentStorageNamespace } from './types/FrontComponentStorageNamespace';
 export type { ElementGeometrySnapshot } from './types/ElementGeometrySnapshot';
 export type { ViewportGeometrySnapshot } from './types/ViewportGeometrySnapshot';
