+++ packages/twenty-front-component-renderer/src/types/FrontComponentStorageScope.ts
@@ -1,0 +1,7 @@
+import { type FrontComponentStorageArea } from 'twenty-sdk/front-component';
+
+import { type FrontComponentStorageNamespace } from '@/types/FrontComponentStorageNamespace';
+
+export type FrontComponentStorageScope = FrontComponentStorageNamespace & {
+  area: FrontComponentStorageArea;
+};
